dal-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic[email]==2.10.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pymysql==1.1.1
cryptography==44.0.0
Faker==33.0.0
